Login page with features:
Username and Password Login: Users should be able to log in using their username and password.

Registration: There should be an option for new users to register with the system.

Password Recovery: In case users forget their password, there should be a mechanism in place for them to recover/reset their password.